feat: add DbContext,ConnectionString

+++ SWP_SchoolMedicalManagementSystem_BussinessProject/Entity/HealthCheckupResults.cs
@@ -11,15 +11,15 @@ namespace SWP_SchoolMedicalManagementSystem_BussinessOject.Entity
         public Guid ScheduleId { get; set; } 
         public StudentHealthCheckupSchedule? HealthCheckupSchedule { get; set; }
         public DateTime CheckupDate { get; set; }
-        public decimal? Height { get; set; } 
-        public decimal? Weight { get; set; } 
+        public float? Height { get; set; } 
+        public float? Weight { get; set; } 
         public string? VisionLeftResult { get; set; }
         public string? VisionRightResult { get; set; }
         public string? HearingLeftResult { get; set; }
         public string? HearingRightResult { get; set; }
-        public int? BloodPressureSys { get; set; } 
-        public int? BloodPressureDia { get; set; } 
-        public int? HeartRate { get; set; } 
+        public float? BloodPressureSys { get; set; } 
+        public float? BloodPressureDia { get; set; } 
+        public float? HeartRate { get; set; } 
         public string? DentalCheckupResult { get; set; }
         public string? OtherResults { get; set; }
         public string? AbnormalSigns { get; set; }
